requirements.txt
Flask==2.1.0
gunicorn==20.1.0

google-cloud==0.34.0
google-cloud-secret-manager==2.1.0